combine popup stories
